Extract gift-wrapping from encryption.ts
